Added btcMine: A reimplementation of mining as a babysitter class that chooses between several miners and chooses the best one. Currently only one miner is implemented - the CPUMiner, which is the original Satoshi client miner, however with the needed CryptoPP functions inlined.
Block: added an updateMerkleTree and a setNonce to enable mining.
Key: Defined vector<unsigned char> as PubKey - to make it explicit when a parameter is a PubKey

BlockChain: added a unconfirmedTransactions method to retrieve all in memory transactions for a new block candidate.
Node: post block added

btcMine:
Miner: as mentioned above, note that the auto detection has not been enabled yet - we postpone this till we have more than one miner.
MinerRPC: SetGenerate, GetGenerate and GetHashesPerSec implemented
CPUMiner: The non assembler version of SHA256::Transform is used for simplicity - the hash/sec is quite modest... (100kHashes on my laptop)

bitcoind: updated to support mining and --gen option