This commit closes #15 - SSL connections are reenabled through additions to the Server Connection and Client classes. The additions are quite trivial mainly following boost examples. Note that the client supports ssl by parsing the http / https protocol. Also, note, that, like in the bitcoind client, no attempts are made for verifying the ssl certificate, hence use at your own risk!.
This commit also patches the ponzicoin example. It now includes a poor mans miner that lets you easily play with the example. In the current setup ponzicoin will generate coins every minute, but typically starts out somewhat faster on a normal machine with coins every 6 seconds. It is great fun to see the coins tick over the screen ;)