Separated out a process method in Block and TransactionFIlter. The Call scheme is : process(tx/blk, peers), and it is called both from the filter, and trough the node.post(tx/blk) via dispatch. The wallet calls node.post(tx) for new transactions and for old transaction not yet in a block. The miner calls node.post(blk) for each generated block, ensuring both network broadcast and wallet notification. The fixes #16
For inbound connections the version returned contained the block height of itself, and hence no new block were requested. This has been changed and hence fixes #14
Further, the SendBufferSize has been increased from 10k to 10*1000k. This was an error in translating the satoshi client options, where I forgot to multiply with 1000