This commit contains two fixes: * To the build system, thanks to G. Maxwell (errors would pop up if you had Qt installed) and for noting the use of shared libraries on Unixes * Threading in the BlockChain, thanks for pointing it out to P. Wuille.
The thread issue could cause an RPC call to fail/the server to die, as there were no protection of the const reading routines of the BlockChain. I analyzed the BlockChain states again.
There are two events that causes the BlockChain state to change:
* new Block
* new Transaction
The latter only changes the Transaction memory pool, whereas the former also changes the database.
Satoshi had utilized the transaction feature of Berkeley DB nicely - all database transactions involved in a new block update (whereas it is just a new block or it also involved a reorganization) are done using TxnBegin and TxnCommit/Abort. This means that the state of the database does not change until the call to TxnCommit. So all simultaneous reads during a reorganization are thread safe and will not access any dirty database state.

Once a new best chain is found or a reorganization is done and we are ready to call TxnCommit, we then update the transaction memory pool. This mean that the actual critical blocks are situated at two spots:
* last third of reorganize
* few lines in the middle of setBestChain

Further, we have, as I mentioned above, also state change due to new Transactions. They will enter through acceptTransaction -> AcceptToMemoryPool. I had to make a separate AcceptToMemoryPool for the new transactions only as the lock I use is not reentrant and we would like to avoid deadlocks. This extra AcceptToMemoryPool is the last place where state changes takes place.

Now that the state changes were isolated it was only a matter of choosing the right way to protect them. I chose boost::unique_lock and boost::shared_lock. They enable a "multiple read, one write" which fit quite well to the way the const interface of the BlockChain is used. Hence the 3 places I mentioned before are protected by a unique_lock, and all public const methods are protected by a shared_lock enabling several of these to be called at once.

One final note: It might seem as there is a possible deadlock from calling connectInputs from CheckMemoryPool from AcceptToMemoryPool from reorganize. However, connectInputs are only called if checkInputs is true, which is only the case for new transactions not for new blocks. Hence there will not be a dead lock.