New application: coinexplorer. Coin explorer uses the Explorer class of the new coinStat library to build a database based on an existing block chain and Node to enable extracting Coins from Addresses. Several RPC commands has also been defined for getting coins from addresses to a general search algorithm
The coinStat Explorer saves its state in a file "explorer.dat" for every 1000 blocks. The format is CDataStream based and for that purpose CDataStream has been extended to support also multimaps.

With the coinexplorer application running you can visit it with a web browser at localhost:8332 and query for Chain Addresses, uint160 addresses and block or transaction hashes in uint256.

To enable an easy web interface the RequestHandler of coinHTTP has been improved to enable handling of "text/plain" content type from web forms.