Update .gitignore and remove three auto-generated files from git
This ignores files generated by cmake and make, and drops three
files from git that are auto-generated by cmake.
It also un-ignores files that no longer exist, as executables are
now generated in bin/, not in the src/ directory.
With these changes, "git status" and "git diff" now worki properly.