Updated Dogecoin according to: http://www.reddit.com/r/dogecoin/comments/2egjqg/dogecoin_core_18_released_mandatory_upgrade/ - not tested!
Added DogeTest, not tested though as the DogeTest network seems rarely operating

Fixed Alerts to not notify unless AppliesTo us…

Changed from lazy purging to minimal purging as the default (to keep indexes from becoming too large) - minimal purging is now exactly the maturity for the chain.

Changed async_write_some to async_write - this could be the culprit for the strange behavior of not sending transactions to peers - simply if the connection was never properly flushed.